requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
python-multipart
langchain
langchain-community
langchain-groq
langchain-text-splitters
pinecone
pypdf
streamlit
python-dotenv
supabase